feat(http): raw/pretty toggle, jq handoff and spooled large bodies (#2157)
The response viewer already pretty-printed, highlighted and folded JSON
bodies. What it could not do was show them any other way, hand one to jq
without a detour, or survive a large one.

- `t` / `http.toggleRawBody` switches between the pretty-printed body and
  the bytes as received. The flag belongs to the view, not the response, so
  history browsing keeps it; highlighting stays on either way.
- Formatting is capped at `PrettyLimit` (2 MiB): past it the body composes
  as plain rows — no indent, no highlight index, no fold scan — with a
  warning row saying so. All three passes are linear in the body size.
- Bodies past `SpoolThreshold` (1 MiB) stream to a spool file while they
  arrive; `Response.Body` keeps only the head. `FullBody`, `BodyReader` and
  `BodyRange` are the explicit ways past it, so captures, the raw-body save
  (now a stream) and `httpdiff` still see everything while the full copy
  lives for one operation instead of for the session.
- The pane states `(showing the first 1.0 MiB of 7.3 MiB — m load more · o
  open as file)`. `m` reads the next window off the spool, `o` opens the
  whole body as an editor tab through `httppane.OpenBodyFileMsg`.
- The history store adopts the spool file into `.ike/http/bodies/` and
  prunes it with its entry; `CleanupSpool` clears the per-process temp
  directory on exit and creating one sweeps what a crash left behind.
- `q` / `http.jqPlayground` opens the jq playground over the shown body, in
  the response pane itself, with the *whole* body as its input.

+++ internal/app/http_save_test.go
@@ -283,7 +283,7 @@ func TestHTTPResponseFileNameWithoutSnapshot(t *testing.T) {
 // TestByteCountLabel: the notification reads in the unit that fits.
 func TestByteCountLabel(t *testing.T) {
 	cases := []struct {
-		n    int
+		n    int64
 		want string
 	}{{0, "0 bytes"}, {512, "512 bytes"}, {2048, "2.0 KiB"}, {3 * 1024 * 1024, "3.0 MiB"}}
 	for _, tc := range cases {
